% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutAugment.R
\name{WATT.Aug.sandwich}
\alias{WATT.Aug.sandwich}
\title{Augmented WATT estimation and sandwich variance (only linear model for outcomes and logistic regression for PS allowed)}
\usage{
WATT.Aug.sandwich(y, z, X, X.out, alpha = 0, epsilon = 0.001, weight = "att")
}
\arguments{
\item{y}{the outcome variable, a numeric vector}

\item{z}{the binary treatment assignment, a vector, values need to be 0 or 1}

\item{X}{predictors (confounders) for the propensity score, a matrix or a vector (if only one covariate)}

\item{X.out}{predictors (confounders) for the two potential outcomes, a matrix or a vector (if only one covariate)}

\item{alpha}{trimming or truncation threshold, if using ATT trimming/truncation}

\item{epsilon}{the variance parameter for the smooth trimming method}

\item{weight}{propensity score weighting methods in weighted ATT estimand, valued from
"att", "trimming", "re-est trimming", "smooth trimming", "truncation",
"overlap", "matching", and "entropy";
The "att" means the conventional ATT estimand;
The "trimming" and "re-est trimming": "re-est trimming" will re-estimate the propensity score on the trimmed sample,
where the "trimming" will not;
For "trimming", "re-est trimming" and "truncation", the alpha argument must be assigned;
For "smooth trimming", the epsilon argument must also be assigned}
}
\description{
Augmented WATT estimation and sandwich variance (only linear model for outcomes and logistic regression for PS allowed)
}
