% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WATTmethods.R
\name{WATT.bootstrap}
\alias{WATT.bootstrap}
\title{Function WATT.bootstrap(): implements the bootstrap variance estimation of the weighting estimator of a WATT estimand}
\usage{
WATT.bootstrap(
  y,
  z,
  X,
  alpha = 0,
  epsilon = 0.001,
  weight = "att",
  trt.SL.library = c("SL.glm"),
  N.boot = 200,
  seed = 509,
  conf.level = 0.95
)
}
\arguments{
\item{N.boot}{number of bootstrap replicates}

\item{seed}{seed for random numbers, used for bootstrap sampling}

\item{conf.level}{level of constructed confidence interval (using normal approximation), with default 0.95
Other parameters are the same in the WATT() function above}
}
\description{
Function WATT.bootstrap(): implements the bootstrap variance estimation of the weighting estimator of a WATT estimand
}
