% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSweightEst.R
\name{WATT.PSW.sandwich}
\alias{WATT.PSW.sandwich}
\title{Sandwich variance estimation for WATT weighting estimator (only allows logistic regression model for PS)}
\usage{
WATT.PSW.sandwich(
  y,
  z,
  X,
  alpha = 0,
  epsilon = 0.001,
  weight = "att",
  conf.level = 0.95
)
}
\arguments{
\item{y}{the outcome variable, a numeric vector}

\item{z}{the binary treatment assignment, a vector, values need to be 0 or 1}

\item{X}{predictors for the propensity score (confounders), a matrix or a vector (if only one covariate)}

\item{alpha}{trimming or truncation threshold, if using ATT trimming/truncation}

\item{epsilon}{the variance parameter for the smooth trimming method}

\item{weight}{propensity score weighting methods in weighted ATT estimand, valued from
"att", "trimming", "re-est trimming", "smooth trimming", "truncation",
"overlap", "matching", and "entropy";
The "att" means the conventional ATT estimand;
The "trimming" and "re-est trimming": "re-est trimming" will re-estimate the propensity score on the trimmed sample,
where the "trimming" will not;
For "trimming", "re-est trimming" and "truncation", the alpha argument must be assigned;
For "smooth trimming", the epsilon argument must also be assigned}

\item{conf.level}{level of constructed confidence interval (using normal approximation), with default 0.95}
}
\description{
Sandwich variance estimation for WATT weighting estimator (only allows logistic regression model for PS)
}
