% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WATTmethods.R
\name{WATT.SumStat}
\alias{WATT.SumStat}
\title{Function WATT.SumStat(): summary statistics in propensity score analysis}
\usage{
WATT.SumStat(
  y,
  z,
  X,
  alpha = 0,
  epsilon = 0.001,
  weight = "att",
  trt.SL.library = c("SL.glm"),
  SumStat = "PS.plot"
)
}
\arguments{
\item{SumStat}{summary statistics need to return, chosen from
               "PS.plot" - the propensity score plots by group
               "cov.bal" - covariate balancing by absolute standard mean difference (by the chosen "weight")
               "all" - return all summary statistics above
Other parameters are the same in the WATT() function above}
}
\description{
Function WATT.SumStat(): summary statistics in propensity score analysis
}
