% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSweightEst.R
\name{WATT.PSW.sandwich}
\alias{WATT.PSW.sandwich}
\title{Sandwich variance estimation for WATT weighting estimator (only allows logistic regression model for PS)}
\usage{
WATT.PSW.sandwich(
  y,
  z,
  X,
  alpha = 0.05,
  epsilon = 0.001,
  weight = "att",
  conf.level = 0.95
)
}
\arguments{
\item{y}{the outcome variable, a numeric vector}

\item{z}{the binary treatment assignment, a vector, values need to be 0 or 1}

\item{X}{predictors for the propensity score (confounders), a matrix or a vector (if only one covariate)}

\item{alpha}{trimming or truncation threshold, if using ATT trimming/truncation}

\item{epsilon}{the variance parameter for the smooth trimming method}

\item{weight}{propensity score weighting methods in weighted ATT estimand, valued from
"att", "trimming", "smooth trimming", "truncation", "overlap", "matching", and "entropy";
The "att" means the conventional ATT estimand;
For "trimming" and "truncation", the `alpha` argument must be assigned; default is 0.05;
For "smooth trimming", the `epsilon` argument must also be assigned; default is 0.001}

\item{conf.level}{level of constructed confidence interval (using normal approximation); default is 0.95}
}
\description{
Sandwich variance estimation for WATT weighting estimator (only allows logistic regression model for PS)
}
